自分の投稿だけ編集・削除可にする #8

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tweet;
+use App\Services\TweetService;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
@@ -14,9 +16,13 @@ class DeleteController extends Controller
      * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int)$request->route('tweetId');
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweetId)) {
+            return new AccessDeniedHttpException();
+        }
+
         $tweet = Tweet::where('id', $tweetId)->firstOrfail();
         $tweet->delete();
         return redirect()
